Move constants with field enums to separate module
Move constants storing reused field enums out of metadata.fields and
to a new metadata.enums module.
